[bouqueau] fix warnings and cosmetics
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3752 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/isomedia/isom_read.c
@@ -1442,8 +1442,8 @@ GF_Err gf_isom_get_sample_for_movie_time(GF_ISOFile *the_file, u32 trackNumber, 
 	//to the media time scale (used by SLConfig) - add the edit start time but stay in
 	//the track TS
 	if (useEdit) {
-		u64 _ts = segStartTime * tsscale;
-		u64 _offset = mediaOffset * tsscale;
+		u64 _ts = (u64)(segStartTime * tsscale);
+		u64 _offset = (u64)(mediaOffset * tsscale);
 
 		(*sample)->DTS += _ts;
 		/*watchout, the sample fetched may be before the first sample in the edit list (when seeking)*/
